adjusted the header files #include statement, moved some functions to merge_external.c
