+++ main.c
@@ -4,95 +4,6 @@
 
 char *write_file_path = "dataset/phase_2.dat";
 
-/**
- * Initialize a MergeManager merger for Phase II
- * Return SUCCESS or FAILURE
- **/
-int init_MergeManager(MergeManager *merger, 
-								int total_sort_runs, int mem_size, int block_size){
-	/* K */
-	int K = total_sort_runs;
-	Record **input_buffers;
-
-	/* size of each buffer - aligned with block_size */
-	int buf_size = get_buf_size(mem_size, block_size, K);
-	if (buf_size <= 0){
-		fprintf(stderr, "buffer size\n");
-		return FAILURE;
-	}
-
-	/* check if the bookeeping structure has enough space */
-	if (!has_enough_mem(mem_size, buf_size, K)){
-		fprintf(stderr, "Not enough memory for Phase II\n");
-		return FAILURE;
-	}
-	
-	/* calloc - heap */
-	if ( !(merger->heap = calloc(K, sizeof(HeapElement))) ){
-		fprintf(stderr, "calloc \n");
-		return FAILURE;
-	}
-	merger->current_heap_size = K;
-	merger->heap_capacity = K;
-
-	/* input_file_numbers */
-	if ( !(merger->input_file_numbers = calloc(K, sizeof(int))) ){
-		fprintf(stderr, "calloc \n");
-		return FAILURE;
-	}
-
-	/* outputFP */
-	if ( !(merger->outputFP = fopen(write_file_path, "wb")) ){
-		fprintf(stderr, "Could not open file %s for writing", write_file_path);
-		return FAILURE;
-	}
-
-	/* output_buffer */
-	if ( !(merger->output_buffer = calloc(1, buf_size)) ){
-		fprintf(stderr, "calloc \n");
-		return FAILURE;
-	}
-
-	merger->current_output_buffer_position = 0;
-	merger->output_buffer_capacity = buf_size;
-
-	/* input_buffers */
-	merger->input_buffers = calloc(K, sizeof(Record*));
-	input_buffers = merger->input_buffers;
-
-	int i;
-	for (i = 0; i < K; i ++){
-		if (! ( input_buffers[i] = calloc(1, buf_size)) ){
-			fprintf(stderr, "calloc \n");
-			return FAILURE;
-		}
-	}
-
-	merger->current_input_file_positions = calloc(K, sizeof(int));
-	if (!merger->current_input_file_positions){
-		fprintf(stderr, "calloc \n");
-		return FAILURE;
-	}
-
-	merger->current_input_buffer_positions = calloc(K, sizeof(int));
-	if (!merger->current_input_buffer_positions){
-		fprintf(stderr, "calloc \n");
-		return FAILURE;
-	}
-
-	merger->total_input_buffer_elements = calloc(K, sizeof(int));
-	if (!merger->total_input_buffer_elements){
-		fprintf(stderr, "calloc \n");
-		return FAILURE;
-	}
-
-	/* file names and input prefix */
-	strcpy(merger->output_file_name, write_file_path);
-	strcpy(merger->input_prefix, INPUT_PREFIX);
-
-	return SUCCESS;
-}
-
 
 int main (int argc, char **argv) {
 	FILE *fp_read;
@@ -112,7 +23,6 @@ int main (int argc, char **argv) {
 	mem_size = get_size_byte(argv[2]);
 	block_size = get_size_byte(argv[3]);	
 	
-	/* param checks */
 	if ((block_size <= 0) || (mem_size <= 0)){
 		fprintf(stderr,"<block size> or <memory size> are invalid.\n");
 		return FAILURE;
